Move directives to layout.module
