+++ alv-portal-ui/src/app/shared/forms/forms.module.ts
@@ -15,13 +15,15 @@ import {
 } from '@ng-bootstrap/ng-bootstrap';
 import { DateParserFormatter } from './input/date-input/date-parser-formatter';
 import { FormsModule as NgFormsModule, ReactiveFormsModule } from '@angular/forms';
+import { IsAuthenticatedDirective } from '../layout/is-authenticated/is-authenticated.directive';
+import { HasAnyAuthorityDirective } from '../layout/has-any-authority/has-any-authority.directive';
 
 @NgModule({
   imports: [
     CommonModule,
     NgFormsModule,
     ReactiveFormsModule,
-    NgbDatepickerModule,
+    NgbDatepickerModule
   ],
   declarations: [
     AddressInputComponent,
@@ -31,7 +33,7 @@ import { FormsModule as NgFormsModule, ReactiveFormsModule } from '@angular/form
     ValidationMessagesComponent,
     DateInputComponent,
     RadioButtonComponent,
-    CheckboxComponent,
+    CheckboxComponent
   ],
   exports: [
     AddressInputComponent,
@@ -41,7 +43,7 @@ import { FormsModule as NgFormsModule, ReactiveFormsModule } from '@angular/form
     ValidationMessagesComponent,
     DateInputComponent,
     RadioButtonComponent,
-    CheckboxComponent,
+    CheckboxComponent
   ],
   providers: [
     NgbPopoverConfig,
